requirements.txt
chromadb
jupyterlab-server
langchain
langchain-chroma
langchain-core
langchain-ollama
langchain_community
langgraph
matplotlib-inline
mkdocs
mkdocs-material
mkdocstrings[python]
nbformat
notebook
numpy
ollama
openai
pandas
pygments
pytest
python-dotenv
six
tqdm
